Remove redundant components and refactor useAssets
